Tests/LibWeb/Text/input/Streams/TransformStream-cancelled-by-writable-side.html

<script src="../include.js"></script>
<script>
    function testTransformStreamReadableStreamCancellation() {
        const transformStream = new TransformStream({
            cancel(reason) {
                println(`TransformStream cancelled due to: ${reason}`);
                return Promise.resolve();
            }
        });
        transformStream.writable.abort('abort called by the writable side.');
    }

    test(() => {
        testTransformStreamReadableStreamCancellation();
    });
</script>
